ingestion/api/requirements.txt
fastapi==0.115.*
uvicorn==0.30.*
kafka-python==2.0.*
pydantic==2.*